Move components into folders and refactor page container components (#3060)
* Move components into folders, remove page components

* Update tests

* Fix test
