--- src/js/edu-benefits/containers/BenefitsSelectionPage.jsx
@@ -1,35 +1,0 @@
-import React from 'react';
-
-import { connect } from 'react-redux';
-
-import BenefitsSelectionFields from '../components/BenefitsSelectionFields';
-import { veteranUpdateField } from '../actions/index';
-
-class BenefitsSelection extends React.Component {
-  render() {
-    const { data, onStateChange } = this.props;
-
-    return (
-      <div className="form-panel">
-        <BenefitsSelectionFields data={data} onStateChange={onStateChange}/>
-      </div>
-    );
-  }
-}
-
-function mapStateToProps(state) {
-  return {
-    data: state.veteran
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    onStateChange(field, update) {
-      dispatch(veteranUpdateField(field, update));
-    }
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(BenefitsSelection);
-export { BenefitsSelection };
